Add Sync tests for RLMSet / MutableSet, RLMArray / List (#7190)

+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -1097,7 +1097,7 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
                 return
             }
 
-            guard case let .int64(sum) = bson else {
+            guard case let .int32(sum) = bson else {
                 XCTFail(error!.localizedDescription)
                 return
             }
